requirements.txt
numpy==2.1.3
matplotlib==3.10.1
cupy-cuda12x==13.4.0